Allow bulk upload of text-only reps Refactor bulk uploads DPAC-204 #resolve
